requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.3
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
python-dotenv==1.1.1
numpy>=1.24.0
rapidfuzz>=3.0.0
torch
